+++ examples/deno/astro.config.mjs
@@ -1,10 +1,9 @@
 import { defineConfig } from 'astro/config';
 
+import deno from '@astrojs/deno';
+
 // https://astro.build/config
 export default defineConfig({
-	vite: {
-		ssr: {
-			noExternal: ['@example/my-component'],
-		},
-	},
+	output: 'server',
+	adapter: deno(),
 });