+++ packages/astro/src/vite-plugin-scripts/index.ts
@@ -1,5 +1,5 @@
 import { ConfigEnv, Plugin as VitePlugin } from 'vite';
-import { AstroConfig, InjectedScriptStage } from '../@types/astro.js';
+import { AstroSettings, InjectedScriptStage } from '../@types/astro.js';
 
 // NOTE: We can't use the virtual "\0" ID convention because we need to
 // inject these as ESM imports into actual code, where they would not
@@ -11,7 +11,7 @@ export const BEFORE_HYDRATION_SCRIPT_ID = `${SCRIPT_ID_PREFIX}${
 export const PAGE_SCRIPT_ID = `${SCRIPT_ID_PREFIX}${'page' as InjectedScriptStage}.js`;
 export const PAGE_SSR_SCRIPT_ID = `${SCRIPT_ID_PREFIX}${'page-ssr' as InjectedScriptStage}.js`;
 
-export default function astroScriptsPlugin({ config }: { config: AstroConfig }): VitePlugin {
+export default function astroScriptsPlugin({ settings }: { settings: AstroSettings }): VitePlugin {
 	let env: ConfigEnv | undefined = undefined;
 	return {
 		name: 'astro:scripts',
@@ -29,19 +29,19 @@ export default function astroScriptsPlugin({ config }: { config: AstroConfig }):
 
 		async load(id) {
 			if (id === BEFORE_HYDRATION_SCRIPT_ID) {
-				return config._ctx.scripts
+				return settings.scripts
 					.filter((s) => s.stage === 'before-hydration')
 					.map((s) => s.content)
 					.join('\n');
 			}
 			if (id === PAGE_SCRIPT_ID) {
-				return config._ctx.scripts
+				return settings.scripts
 					.filter((s) => s.stage === 'page')
 					.map((s) => s.content)
 					.join('\n');
 			}
 			if (id === PAGE_SSR_SCRIPT_ID) {
-				return config._ctx.scripts
+				return settings.scripts
 					.filter((s) => s.stage === 'page-ssr')
 					.map((s) => s.content)
 					.join('\n');
@@ -49,7 +49,7 @@ export default function astroScriptsPlugin({ config }: { config: AstroConfig }):
 			return null;
 		},
 		buildStart(options) {
-			const hasHydrationScripts = config._ctx.scripts.some((s) => s.stage === 'before-hydration');
+			const hasHydrationScripts = settings.scripts.some((s) => s.stage === 'before-hydration');
 			if (hasHydrationScripts && env?.command === 'build' && !env?.ssrBuild) {
 				this.emitFile({
 					type: 'chunk',