+++ packages/integrations/image/src/loaders/index.ts
@@ -19,7 +19,9 @@ export type ColorDefinition =
 	| NamedColor
 	| `#${string}`
 	| `rgb(${number}, ${number}, ${number})`
-	| `rgb(${number},${number},${number})`;
+	| `rgb(${number},${number},${number})`
+	| `rgba(${number}, ${number}, ${number}, ${number})`
+	| `rgba(${number},${number},${number},${number})`;
 
 export type CropFit = 'cover' | 'contain' | 'fill' | 'inside' | 'outside';